Separate cocoalibs into separate library files. Some of these do not support iOS, so keeping them separate allows branching imports. Separate quartz, coregraphics, and coretext.  Quartz seems misnamed all over. Is actually the display functionality, not font or image related like some modules seem to think. Many of the "Quartz" functionality comes form CoreGraphics, just re-exported. Update these namings. May rename quartz namings in the future.
