+++ pyglet/libs/darwin/cocoapy/__init__.py
@@ -28,12 +28,21 @@
 # LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-from contextlib import contextmanager
 
 from .runtime import objc, send_message, send_super, AutoReleasePool, ObjCBlock
 from .runtime import get_selector
 from .runtime import ObjCClass, ObjCInstance, ObjCSubclass
 
 from .cocoatypes import *
-from .cocoalibs import *
 from .cocoahelpers import *
+
+from .lib_corefoundation import *
+from .lib_coregraphics import *
+from .lib_coretext import *
+
+if sys.platform == 'darwin':
+    from .lib_appkit import *
+    from .lib_cocoa import *
+    from .lib_quartz import *
+elif sys.platform == 'ios':
+    from .lib_uikit import *
