+++ pyglet/libs/darwin/cocoapy/lib_uikit.py
@@ -1,0 +1,7 @@
+"""UIKit for iOS."""
+from __future__ import annotations
+
+import pyglet.lib
+
+uikit = pyglet.lib.load_library(framework='UIKit')
+quartzcore = pyglet.lib.load_library(framework='QuartzCore')
