+++ pyglet/libs/darwin/lib_agl.py
@@ -1,12 +1,19 @@
 from __future__ import annotations
 
+import sys
 from typing import Any, Callable, Sequence
 
 import pyglet.lib
 from pyglet.graphics.api.gl.lib import decorate_function, missing_function
 
-gl_lib = pyglet.lib.load_library(framework='OpenGL')
-agl_lib = pyglet.lib.load_library(framework='AGL')
+if sys.platform == 'ios':
+    # iOS only.
+    gl_lib = pyglet.lib.load_library(framework='OpenGLES')
+    agl_lib = None
+else:
+    # Desktop Only.
+    gl_lib = pyglet.lib.load_library(framework='OpenGL')
+    agl_lib = pyglet.lib.load_library(framework='AGL')
 
 
 def link_GL(name: str, restype: Any, argtypes: Any, requires: str | None = None,  # noqa: N802
@@ -23,6 +30,8 @@ def link_GL(name: str, restype: Any, argtypes: Any, requires: str | None = None,
 
 def link_AGL(name: str, restype: Any, argtypes: Any, requires: str | None = None,  # noqa: N802
              suggestions: Sequence[str] | None = None) -> Callable[..., Any]:
+    if agl_lib is None:
+        return missing_function(name, requires, suggestions)
     try:
         func = getattr(agl_lib, name)
         func.restype = restype
